% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPRegions-package.R
\docType{package}
\name{ChIPRegions-package}
\alias{ChIPRegions}
\alias{ChIPRegions-package}
\title{ChIPRegions: Modelling Regional Bias in ChIP-seq}
\description{
This package contains the functions necessary to transform a collection of GRanges objects into a data structure that can be fit with a HMM.
}
\keyword{internal}
