% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromGet.R
\name{ChromGet}
\alias{ChromGet}
\title{ChromGet}
\usage{
ChromGet(UnifiedSequence, MaxGap = NULL, MinPeaks = 2)
}
\arguments{
\item{UnifiedSequence}{Output of Unify function, Grange object}

\item{MaxGap}{Numeric, maximum gap allowed between peak unions within the same subsequence}

\item{MinPeaks}{Numeric, the minimum number of peaks required to form a valid subsequence}
}
\description{
Splits the peak union sequence into sub sequences for the HMM, each sub sequence will be considered an
I.I.D sample from the same hidden markov process. This is necessary to remove state transitions
without biological meaning from the model, such as between the last peak union on chromosome N and the first of chromosome N+1.
By default, sub sequences are only generated for each chromosome, but a maximum gap "MaxGap" between peaks can be
specified such that the chain is assumed to "start afresh" if the distance to the next peak is > MaxGap.
}
