% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegions.R
\name{GetRegions}
\alias{GetRegions}
\title{Splits genome into contiguous regions defined by state annotations}
\usage{
GetRegions(ChromSplit, Viterbi, UnifiedSequence)
}
\arguments{
\item{ChromSplit}{A list of sequences of TRUE/FALSE tuples. Obtained after splitting the Peak Union Sequence by chromosome/genomic features.}

\item{Viterbi}{A list of Viterbi sequences. Each Viterbi sequence corresponds to a sequence in ChromSplit.}

\item{UnifiedSequence}{a GRanges object containing the unified intervals from which ChromSplit and Viterbi are derived.}
}
\description{
Splits genome into contiguous regions defined by state annotations
}
