% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreRegions.R
\name{ScoreRegions}
\alias{ScoreRegions}
\title{ScoreRegions}
\usage{
ScoreRegions(
  UnionSequence,
  Viterbi,
  Bedlist,
  Scorename = "SignalValue",
  ReferenceState = 1,
  MinSample = 20
)
}
\arguments{
\item{UnionSequence}{Output of Unify function}

\item{Viterbi}{A list of Viterbi sequences. Each Viterbi sequence corresponds to a sequence in ChromSplit.}

\item{Bedlist}{A list of GRanges objects used to create UnionSequence}

\item{Scorename}{Character value, the name of the metadata column to be compared across regions}

\item{ReferenceState}{Numeric value, the HMM state to be used as the baseline}

\item{MinSample}{Numeric value, overlap patterns will removed from the comparison if any state contains less than MinSample observations}
}
\description{
Computes a regional signal score. Computing this score requires stratifying across each overlap pattern
and can only be computed for overlap patterns present within each state. The minimum sample size for each state can be set with MinSample.
The score is reported as a ratio to a chosen reference state, changing the reference state does not influence the ratio between the scores of any two states.
}
